# backend/src/docs/openapi.snapshot.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GitNest API",
    "version": "1.0.0",
    "description": "Schema-driven API contract for GitNest MERN services."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "GitNest API server"
    }
  ],
...
  "components": {
    "schemas": {
...
      "ErrorEnvelope": {
        "type": "object",
        "additionalProperties": true,
        "properties": {
          "success": {
            "const": false
          },
          "status": {
            "type": "string",
            "enum": [
              "fail",
              "error"
            ]
          },
          "statusCode": {
            "type": "integer",
            "minimum": 400
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true,
              "properties": {
                "field": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "message"
              ]
            }
          },
          "requestId": {
            "type": [
              "string",
              "null"
            ]
          },
          "timestamp": {
            "type": "string"
          }
        },
        "required": [
          "success",
          "code",
          "message",
          "errors",
          "requestId"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/v1/repositories": {
      "post": {
        "tags": [
          "Repositories"
        ],
        "summary": "Create a repository",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 500
                  },
                  "visibility": {
                    "type": "string",
                    "enum": [
                      "public",
                      "private"
                    ]
                  },
                  "language": {
                    "type": "string",
                    "maxLength": 50
                  },
                  "topics": {
                    "type": "array",
                    "maxItems": 20,
                    "items": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 50
                    }
                  },
                  "defaultBranch": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "201 response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "properties": {
                    "success": {
                      "const": true
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "success"
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "additionalProperties": true,
                      "properties": {
                        "_id": {
                          "anyOf": [
                            {
                              "type": "string",
                              "pattern": "^[a-fA-F0-9]{24}$"
                            },
                            {
                              "type": "string"
                            }
                          ]
                        },
                        "name": {
                          "type": "string"
                        },
                        "owner": {},
                        "description": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "visibility": {
                          "type": "string",
                          "enum": [
                            "public",
                            "private"
                          ]
                        },
                        "stars": {
                          "type": "array"
                        },
                        "forks": {
                          "type": "array"
                        },
                        "forkedFrom": {},
                        "defaultBranch": {
                          "type": "string"
                        },
                        "language": {
                          "type": "string"
                        },
                        "topics": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    },
                    "requestId": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Authentication error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  }
}